# vsr: state sync performance improvements