### Summary

This PR improves the performance of state sync (specifically, the syncing of table blocks) using non-intrusive optimizations and configuration changes. Using some simple observations, we were able to improve the time that it takes to sync a ~86G data file from **_13095 seconds (~218 minutes) → 869s (~14.5 minutes)_**. 

These numbers are from running the following simple experiment on a _3-node TB cluster on t2.large AWS instances_:

1. Start R1, R2, but not R0
2. Run tigerbeetle benchmark with `transfer_count=150,000,000`, wait for it to finish
3. Start R0, measure time taken to sync all the content with R1 & R2's help

### Optimizations

Here are the optimizations that we implemented:

1. https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2286193309 explains why we removed this optimization: ~**Commit pipeline (and thereby compaction) paused till content is synced**: The motivation behind pausing the commit pipeline (and thereby compaction) is that we observed a lot of `error(grid): 0: invalid_checksum: expected address=X checksum=Y, found address=0 checksum=0` while table blocks were being synced. This is because compaction tries to read index/data blocks, finds them empty, and queues the block(s) for repair. During state sync, reading a block (and incurring extraneous IO) is a rather roundabout way to to discover that a block needs to synced, when in fact [sync_enqueue_tables](https://github.com/tigerbeetle/tigerbeetle/blob/918c25070e3b479fec561ff00bd1f35c0d6df2a4/src/vsr/replica.zig#L9111) is already doing the same without incurring IO.~ 
2. **Proactive request_blocks**: Inspired by  [this optimization that we have for request_prepares](https://github.com/tigerbeetle/tigerbeetle/blob/c88e34235719a45a8a3f3347f35d9d9d68baef49/src/vsr/replica.zig#L9654-L9657)), we now [proactively send request_blocks](https://github.com/tigerbeetle/tigerbeetle/commit/d14a1f986f12e721742c0734cad6f2a8aea55e01) before the next timeout, in case the syncing replica has enough write iops.
3. **Increasing grid_missing_tables_max from 3 → 6**: Increases the # of tables that are concurrently synced

<details>
<summary>Here is the performance improvement that each of the aforementioned optimizations brings:</summary>


#### No optimizations

```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "sync: ops\|sync_enqueue_tables\|lagging"
warning(replica): 0: on_start_view: start sync; lagging behind cluster (op_prepare_max=999 quorum_head=18420)
info(replica): 0: sync: ops=0..18271
info(replica): 0: sync_enqueue_tables: all tables synced (commit=0..18271) in 13095s
```

#### Compaction Freeze: 

```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "sync: ops\|sync_enqueue_tables: \|lagging"
warning(replica): 0: on_start_view: start sync; lagging behind cluster (op_prepare_max=999 quorum_head=18420)
info(replica): 0: sync: ops=0..18271
info(replica): 0: sync_enqueue_tables: all tables synced (commit=0..18271) in 11034s
```

#### Compaction Freeze + Proactive request_blocks

```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "sync: ops\|sync_enqueue_tables: \|lagging"
warning(replica): 0: on_start_view: start sync; lagging behind cluster (op_prepare_max=999 quorum_head=18420)
info(replica): 0: sync: ops=0..18271
info(replica): 0: sync_enqueue_tables: all tables synced (commit=0..18271) in 1054s
```

#### Compaction Freeze + Proactive request_blocks + grid_missing_tables_max: 3 → 6

```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "sync: ops\|sync_enqueue_tables\|lagging"
warning(replica): 0: on_start_view: start sync; lagging behind cluster (op_prepare_max=999 quorum_head=18420)
info(replica): 0: sync: ops=0..18271
info(replica): 0: sync_enqueue_tables: all tables synced (commit=0..18271) in 869s
```
</details>