@chaitanyabhandari 2024-08-13
From https://github.com/tigerbeetle/tigerbeetle/actions/runs/10362913189/job/28685691012?pr=2207, realized that a few replica tests which test state sync assert that the syncing replica is also able to commit while tables are being synced. We changed this in https://github.com/tigerbeetle/tigerbeetle/pull/2207/commits/971090d9cd790abdd52be75306b832661325717f, wherein we pause then commit pipeline before tables are synced. Before changing all replica tests to reflect this new behaviour, I want to do one more experiment, where we _don't_ pause the commit pipeline, but we do apply the other two optimizations _(Proactive request_blocks + grid_missing_tables_max: 3 → 6)_. If the effect of pausing commit pipeline is negligible in the presence of the other two optimizations, we can get away with not changing this behavior at all.