@chaitanyabhandari 2024-08-13
Re https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2285271858: Hunch confirmed! The benefits of pausing the commit pipeline are indeed diminished by the other two optimizations. We're still able to sync in **_806s._** So, we can get continue to embrace the concurrency here! I will remove https://github.com/tigerbeetle/tigerbeetle/pull/2207/commits/971090d9cd790abdd52be75306b832661325717f.

```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "sync: ops\|sync_enqueue_tables: \|lagging"
info(replica): 0: sync: ops=0..18271
warning(replica): 0: on_repair_sync_timeout: request sync; lagging behind cluster (op_head=18320 commit_min=18239 commit_max=18320 commit_stage=idle)
info(replica): 0: sync_enqueue_tables: all tables synced (commit=0..18271) in 806s
```
cc: @matklad @jorangreef