@chaitanyabhandari 2024-08-13
To get a feel of whether we're close to line rate (the global maxima for the performance we can achieve), we compare our performance with rsync. cc: @jorangreef 

#### Result

To sync **_~52.5G_** worth of data blocks, state sync takes **_806 seconds_** (https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2286193309), while rsync takes **_628.89 seconds_**.

#### Details 

In the process of syncing a ~86G data file, the syncing replica repaired a total of 107432 blocks, i.e. ~52.5G worth of blocks. This takes **_806 seconds_** (https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2286193309).
```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "on_block" | wc -l
107432
```

rsync-ing the ~56G data file from R1 → R2, where R2 doesn't have any data file takes **_640 seconds_** (628.89 seconds if we extrapolate the bandwidth rsync gets to ~52.5G):
```C
ubuntu@ip-172-31-3-127:~/tigerbeetle$ rsync -avmPz -e "ssh -i chaitanyabhandari.pem" data.tigerbeetle ubuntu@172.31.3.11:/home/ubuntu/tigerbeetle/data.tigerbeetle
building file list ...
1 file to consider
data.tigerbeetle
 56,074,723,328 100%   83.48MB/s    0:10:40 (xfr#1, to-chk=0/1)

sent 17,954,925,730 bytes  received 35 bytes  27,988,972.35 bytes/sec
total size is 56,074,723,328  speedup is 3.12
```