@chaitanyabhandari 2024-08-20
To get a feel of whether we're close to line rate (the global maxima for the performance we can achieve), we compare our performance with iperf:



### Result

To sync **_~52.5G_** worth of data blocks, state sync takes **_806 seconds_** (https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2286193309), while rsync takes **_701 seconds_**. 

### Machine Specifications

These experiments were conducted on [t2.large](https://aws.amazon.com/ec2/instance-types/t2/) instances with 8G RAM and 2 cores:
```C
ubuntu@ip-172-31-3-127:~/tigerbeetle$ free -mh
               total        used        free      shared  buff/cache   available
Mem:           7.8Gi       3.6Gi       167Mi       940Ki       4.3Gi       4.2Gi
Swap:             0B          0B          0B

ubuntu@ip-172-31-3-127:~/tigerbeetle$ lscpu
Architecture:             x86_64
  CPU op-mode(s):         32-bit, 64-bit
  Address sizes:          46 bits physical, 48 bits
                          virtual
  Byte Order:             Little Endian
CPU(s):                   2
  On-line CPU(s) list:    0,1
Vendor ID:                GenuineIntel
  Model name:             Intel(R) Xeon(R) CPU E5-26
                          86 v4 @ 2.30GHz
    CPU family:           6
    Model:                79
    Thread(s) per core:   1
    Core(s) per socket:   2
```

### Details 

In the process of syncing a ~86G data file, the syncing replica repaired a total of 107432 blocks, i.e. ~52.5G worth of blocks. This takes **_806 seconds_** (https://github.com/tigerbeetle/tigerbeetle/pull/2207#issuecomment-2286193309).
```C
ubuntu@ip-172-31-3-11:~/tigerbeetle$ cat replica0.vopr | grep "on_block" | wc -l
107432
```
Sending a ~52.2G of data using iperf takes **_701 seconds_**:
```C
- - - - - - - - - - - - - - - - - - - - - - - - -
[ ID] Interval           Transfer     Bitrate         Retr
[  5]   0.00-701.00 sec  52.2 GBytes   640 Mbits/sec  126             sender
        Sent 52.2 GByte / 52.2 GByte (100%) of data.tigerbeetle
[  5]   0.00-701.00 sec  52.2 GBytes   640 Mbits/sec                  receiver

iperf Done.
```